+++ CHANGELOG.md
@@ -20,3 +20,7 @@
 #### YOLO for ROS2 추가
 - `src/yolo_percepriton/` : custom msg 패키지
 - `src/yolov11_ros/` : YOLO + ROS2 패키지
+- `src/turtlebot3_agent/turtlebot3_agent/tools/turtlebot3.py`
+    - Subscribe to /detection_results (DetectionArray)
+    - Parse the detection info (label, confidence, bounding_box etc.)
+    - Provide a yolo_tool() function